Refactor error types
