+++ p256k1/src/ecdsa.rs
@@ -1,4 +1,4 @@
-use bs58::{self, decode::Error as DecodeError};
+use bs58;
 use std::array::TryFromSliceError;
 
 use crate::_rename::{
@@ -8,24 +8,8 @@ use crate::_rename::{
 };
 use crate::bindings::{secp256k1_ecdsa_signature, secp256k1_pubkey, SECP256K1_EC_COMPRESSED};
 use crate::context::Context;
+use crate::errors::{Base58Error, ConversionError};
 use crate::scalar::Scalar;
-
-/// Re-export of crate `bs58`'s decode error
-pub type Base58DecodeError = DecodeError;
-
-#[derive(Debug, Clone)]
-/// Base58-related errors
-pub enum Base58Error {
-    /// Error decoding
-    Decode(Base58DecodeError),
-}
-
-/// Errors when converting scalars
-#[derive(Debug, Clone)]
-pub enum ConversionError {
-    /// Error converting a base58 string to bytes
-    Base58(Base58Error),
-}
 
 #[derive(Debug, Clone)]
 /// Errors in ECDSA signature operations
