+++ p256k1/src/point.rs
@@ -1,4 +1,4 @@
-use bs58::{self, decode::Error as DecodeError, encode::Error as EncodeError};
+use bs58;
 use core::{
     cmp::{Eq, PartialEq},
     convert::{From, TryFrom},
@@ -18,6 +18,7 @@ use crate::{
         secp256k1_callback, secp256k1_ecmult_multi_callback, secp256k1_fe, secp256k1_ge,
         secp256k1_gej, secp256k1_scalar, SECP256K1_TAG_PUBKEY_EVEN, SECP256K1_TAG_PUBKEY_ODD,
     },
+    errors::{Base58Error, ConversionError},
     group::secp256k1_ge_set_gej,
 };
 
@@ -61,33 +62,6 @@ pub const N: [u8; 32] = [
     0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFE,
     0xBA, 0xAE, 0xDC, 0xE6, 0xAF, 0x48, 0xA0, 0x3B, 0xBF, 0xD2, 0x5E, 0x8C, 0xD0, 0x36, 0x41, 0x41,
 ];
-
-/// Re-export of crate `bs58`'s decode error
-pub type Base58DecodeError = DecodeError;
-/// Re-export of crate `bs58`'s encode error
-pub type Base58EncodeError = EncodeError;
-
-#[derive(Debug, Clone)]
-/// Base58-related errors
-pub enum Base58Error {
-    /// Error decoding
-    Decode(Base58DecodeError),
-    /// Error encoding
-    Encode(Base58EncodeError),
-}
-
-#[derive(Debug, Clone)]
-/// Errors when converting points
-pub enum ConversionError {
-    /// Error decompressing a point into a field element
-    BadFieldElement,
-    /// Error decompressing a point into a group element
-    BadGroupElement,
-    /// Error converting a byte slice into Compressed
-    WrongNumberOfBytes(usize),
-    /// Error converting a base58 string to bytes
-    Base58(Base58Error),
-}
 
 #[derive(Debug, Clone)]
 /// Errors in point operations
