+++ p256k1/src/scalar.rs
@@ -1,5 +1,5 @@
 use bitvec::prelude::*;
-use bs58::{self, decode::Error as DecodeError, encode::Error as EncodeError};
+use bs58;
 use core::{
     cmp::{Eq, PartialEq},
     convert::{From, TryFrom},
@@ -19,28 +19,7 @@ use crate::bindings::secp256k1_scalar;
 
 use crate::point::Point;
 
-/// Re-export of crate `bs58`'s decode error
-pub type Base58DecodeError = DecodeError;
-/// Re-export of crate `bs58`'s encode error
-pub type Base58EncodeError = EncodeError;
-
-#[derive(Debug, Clone)]
-/// Base58-related errors
-pub enum Base58Error {
-    /// Error decoding
-    Decode(Base58DecodeError),
-    /// Error encoding
-    Encode(Base58EncodeError),
-}
-
-#[derive(Debug, Clone)]
-/// Errors when converting scalars
-pub enum ConversionError {
-    /// Error converting a byte slice into Scalar
-    WrongNumberOfBytes(usize),
-    /// Error converting a base58 string to bytes
-    Base58(Base58Error),
-}
+use crate::errors::{Base58Error, ConversionError};
 
 #[derive(Debug, Clone)]
 /// Errors in scalar operations
